// src/lib/types.ts
import Stripe from "stripe";
import { Json } from "../../database.types";

export type UserDetails = {
  id: string;
  first_name: string;

  last_name: string;
  full_name?: string;
  avatar_url?: string;
  billing_address?: Stripe.Address;
  payment_methods?: Stripe.PaymentMethod[Stripe.PaymentMethod.Type];
};
export type Product = {
  id: string;
  active?: boolean;
  name?: string;
  description?: string;
  image?: string;
  metadata?: Stripe.Metadata;
};

export type Price = {
  id: string;
  product_id?: string;
  active?: boolean;
  description: string;
  unit_amount?: number;
  currency?: string;
  type?: Stripe.Price.Type;
  interval?: Stripe.Price.Recurring.Interval;
  interval_count?: number;
  trial_period_days?: number | null;
  metadata?: Stripe.Metadata;
  products?: Product;
};

export type Subscription = {
  id: string;
  user_id: string;
  status?: Stripe.Subscription.Status;
  metadata?: Stripe.Metadata;
  price_id?: string;
  quantity?: string;
  cancel_at_period_end?: boolean;
  created: string;
  current_period_start: string;
  current_period_end: string;
  ended_at?: string;
  cancel_at?: string;
  canceled_at?: string;
  trial_start?: string;
  trial_end?: string;
  pricec?: Price;
};

export type User = {
  avatar_url: string | null;
  billing_address: Json | null;
  full_name: string | null;
  id: string;
  payment_method: Json | null;
};

export type ModalProps = {
  modalUpOpen?: boolean;
  setModalUpOpen: React.Dispatch<React.SetStateAction<boolean>>;
  modalInOpen?: boolean;
  setModalInOpen: React.Dispatch<React.SetStateAction<boolean>>;
};

export type Song = {
  author: string | null;
  created_at: string;
  id: number;
  image_path: string | null;
  song_path: string | null;
  title: string | null;
  user_id: string | null;
};